folder refactoring, use grid system
